v0.5.1
Redacted extractNS in include/shared.js to handle URLs that also
specify a port for the server. Added comments.

ExtractNS gives correct output with the following inputs:

http://google.it:20/aa
http://google.it:20/
http://google.it:20
google.it:20/aa
google.it:20/
google.it:20

http://google.it/aa
http://google.it/
http://google.it
google.it/aa
google.it/
google.it

Note, the test suite gives complete path coverage of the function at
issue.

Signed-off-by: tomas <tomasbortoli@gmail.com>