v0.6.1
Fix of a minor bug in the ignore-this-black-listed-site function.

The issue laid on the fact that the when the user asked to ignore a
certain site, the code was adding the target name-server to the
ignored-sites-list however if the aforesaid name-server was a >=3rd
level domain, then the user could not reach the site even after the
ignore, because the block logic blocks a certain request by checking the
nameserver from the 2nd level domain.

So for example:
1 - user wants to visit aaa.bbb.com
2 - suppose bbb.com and eventually also aaa.bbb.com in black-list
3 - therefore the user gets blocked
4 - user explicitly ignore the block
5 - Fraud Blocker puts aaa.bbb.com in ignored list
6 - User still can't reach aaa.bbb.com because bbb.com isn't in the
ignored list

Fix: at the 5th point, add bbb.com to the ignore list besides
aaa.bbb.com

Signed-off-by: tomas <tomasbortoli@gmail.com>