v0.6.2
Fix of a malfunction introduced in v0.6.0. The issue laid on the fact
that if a user clicked the popup to ignore the block on a black-listed
site and then clicked again the ignore button, he would get redirected
to a non-existent page on the present web-ext. To fix, the code now
checks whether the page is at a "moz-extension" page or not when
receiving an "ignore" msg before redirecting.

Furthermore, added check to prevent other extensions (potentially
malicious) to send messages to Fraud Blocker to instruct him to do
things like redirect the user to a malicious page, ignore black listed
sites, or give feedback for sites. Now the trusted messages are only
those coming from the same source (Fraud Blocker).

But anyway, this could not really turn into a security issue as
installed extensions already have the capability of opening arbitrary,
new URLs.

Added also check to whitelist Firefox's internal "about" protocol

Signed-off-by: tomas <tomasbortoli@gmail.com>