v0.6.3
This commit derives from the testing of the web-ext in Google Chrome. In
the aforesaid browser I observed a malfunction derived from the fact
that the getStatus function returns "black" if the current page is
black-listed or if it is the blocked-page. Therefore if a user hits a
black-listed page:

- the Chrome browser gets redirected by the handler in background.js to
the blocked-page

- then getStatus is triggered again recursively and with the old
behavior the placeholder blocked-page is detected as black-listed and
therefore redirected again to something like
chrome-extension://...?chrome-extension...?black-listed-url while it
should be chrome-extension://...?black-listed-url

Although in Firefox this doesn't happen (as requests to local web-ext
are treated differently than requests to external systems and therefore
not caught by the "onBeforeRequest" handler), it is a good idea to
differentiate the two things. Therefore,

include/messaging.js: return msg.msg="blocked" instead of "black" if the
current page is the blocked-page placeholder.

popup/guiHandler.js: adapt to the new behavior of include/messaging.js

Signed-off-by: tomas <tomasbortoli@gmail.com>